Moved filter components into TableFilter folder
