+++ react/src/components/Table/TableFilter/GlobalFilters.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useAsyncDebounce } from 'react-table';
 import styled from 'styled-components/macro';
-import { Flex, Input } from '../index';
+import { Flex, Input } from '../..';
 
 interface GlobalFilterProps {
     filter: any;
